[#1] move. Command 패키지 application 내부로 이동
