+++ src/main/java/com/baedal/owner/adapter/in/web/mapper/OwnerWebMapper.java
@@ -2,7 +2,7 @@ package com.baedal.owner.adapter.in.web.mapper;
 
 import org.mapstruct.Mapper;
 
-import com.baedal.owner.adapter.in.web.dto.command.LoginCommand;
+import com.baedal.owner.application.command.LoginCommand;
 import com.baedal.owner.adapter.in.web.dto.request.LoginRequest;
 import com.baedal.owner.adapter.in.web.dto.response.LoginResponse;
