+++ src/main/java/com/baedal/owner/application/mapper/OwnerApplicationMapper.java
@@ -3,7 +3,7 @@ package com.baedal.owner.application.mapper;
 import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
 
-import com.baedal.owner.adapter.in.web.dto.command.LoginCommand;
+import com.baedal.owner.application.command.LoginCommand;
 import com.baedal.owner.domain.model.Owner;
 
 @Mapper(componentModel = "spring")
