+++ src/main/java/com/baedal/owner/application/port/in/OwnerUseCase.java
@@ -1,6 +1,6 @@
 package com.baedal.owner.application.port.in;
 
-import com.baedal.owner.adapter.in.web.dto.command.LoginCommand;
+import com.baedal.owner.application.command.LoginCommand;
 
 public interface OwnerUseCase {
 	LoginCommand.Response login(LoginCommand.Request req);
