+++ src/main/java/com/baedal/owner/application/service/OwnerService.java
@@ -2,7 +2,7 @@ package com.baedal.owner.application.service;
 
 import org.springframework.stereotype.Component;
 
-import com.baedal.owner.adapter.in.web.dto.command.LoginCommand;
+import com.baedal.owner.application.command.LoginCommand;
 import com.baedal.owner.application.mapper.OwnerApplicationMapper;
 import com.baedal.owner.application.port.in.OwnerUseCase;
 import com.baedal.owner.application.port.out.OwnerRepositoryPort;
